Create BaseFileFinder and TestingResultFileFinder classes
The functionality of the CoverageFileFinder class has been
moved into the new BaseFileFinder class that the CoverageFileFinder
and TestingResultFileFinder inherit from.

The goal is to have the two finders share the same functionality for
finding files whilst having different default patterns to look for.
